feat(cosmos): support `InstallBundle` and `SendChunk`

+++ golang/cosmos/x/swingset/types/msgs_test.go
@@ -151,7 +151,7 @@ func TestInstallBundle_ValidateBasic(t *testing.T) {
 			msg: &MsgInstallBundle{
 				Submitter:        addr,
 				CompressedBundle: []byte{1, 2, 3},
-				UncompressedSize: bundleUncompressedSizeLimit - 1,
+				UncompressedSize: int64(bundleUncompressedSizeLimit - 1),
 			},
 		},
 
@@ -160,7 +160,7 @@ func TestInstallBundle_ValidateBasic(t *testing.T) {
 			msg: &MsgInstallBundle{
 				Submitter:        addr,
 				CompressedBundle: []byte{1, 2, 3},
-				UncompressedSize: bundleUncompressedSizeLimit,
+				UncompressedSize: int64(bundleUncompressedSizeLimit),
 			},
 			shouldErr: true,
 		},
